refactor(i18n): DOM依存i18n関数をi18n-dom.tsに分離し副作用を除去
