+++ entrypoints/options/main.ts
@@ -1,6 +1,9 @@
 /**
  * Options/Dashboard entry point - imports from src/dashboard/
  */
-import '../../src/utils/i18n.js';
+import { applyI18n, setHtmlLangAndDir } from '../../src/utils/i18n-dom.js';
 import '../../src/dashboard/dashboard.js';
 import '../../src/dashboard/main.js';
+
+setHtmlLangAndDir();
+applyI18n();
