+++ src/popup/onboardingWizard.ts
@@ -2,7 +2,8 @@ import { getSettings, saveSettings } from '../utils/storage.js';
 import { StorageKeys } from '../utils/storage/types.js';
 import { focusTrapManager } from './utils/focusTrap.js';
 import { logError, ErrorCode } from '../utils/logger.js';
-import { applyI18n, getMessage } from '../utils/i18n.js';
+import { getMessage } from '../utils/i18n.js';
+import { applyI18n } from '../utils/i18n-dom.js';
 
 export type WizardType = 'obsidian' | 'sqlite' | 'minimal';
